Refactor Navbar component and styles; replace Dashboard Navbar with new Navbar structure
