+++ exchange-admin/src/App.css
@@ -109,7 +109,7 @@
 }
 
 table {
-  width: 800px;
+  width: 90vw;
   height: 100%;
   border-collapse: collapse;
   overflow: hidden;
