+++ exchange-admin/src/Components/Dashboard/Dashboard.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navbar from './Navbar'
+import Navbar from '../Navbar/Navbar'
 import './Dashboard.css'
 import logo from '../Assets/logo.png'
